Adding authorize for all methods

+++ modules/RolePermissions/Http/Controllers/RolePermissionsController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use Category\Responses\AjaxResponses;
 use RolePermissions\Http\Requests\RoleRequest;
 use RolePermissions\Http\Requests\RoleUpdateRequest;
+use RolePermissions\Models\Role;
 use RolePermissions\Repositories\PermissionRepo;
 use RolePermissions\Repositories\RoleRepo;
 
@@ -23,6 +24,7 @@ class RolePermissionsController extends Controller
 
     public function index()
     {
+        $this->authorize('view', Role::class);
         $roles = $this->repository_role->all();
         $permissions = $this->repository_permission->all();
         return view('RolePermissions::index', compact('roles', 'permissions'));
@@ -30,11 +32,14 @@ class RolePermissionsController extends Controller
 
     public function store(RoleRequest $request)
     {
-        return $this->repository_role->create($request);
+        $this->authorize('create', Role::class);
+        $this->repository_role->create($request);
+        return redirect(route('RolePermissions::index'));
     }
 
     public function edit($roleId)
     {
+        $this->authorize('edit', Role::class);
         $role = $this->repository_role->findById($roleId);
         $permissions = $this->repository_permission->all();
         return view('RolePermissions::edit', compact('role', 'permissions'));
@@ -42,12 +47,14 @@ class RolePermissionsController extends Controller
 
     public function update(RoleUpdateRequest $request, $id)
     {
+        $this->authorize('edit', Role::class);
         $this->repository_role->update($id, $request);
         return redirect(route('role-permissions.index'));
     }
 
     public function destroy($roleId)
     {
+        $this->authorize('delete', Role::class);
         $this->repository_role->delete($roleId);
         return AjaxResponses::SuccessResponse();
     }
